requirements.txt
fastapi==0.118.2
gigachat==0.1.43
python-dotenv==1.1.1
python-multipart==0.0.20
torch==2.9.1
torchvision==0.24.1
ultralytics==8.3.237
ultralytics-thop==2.0.18
uvicorn==0.37.0
